+++ src/kwja/datamodule/datasets/__init__.py
@@ -1,9 +1,9 @@
-from .char_dataset import CharDataset
-from .char_inference_dataset import CharInferenceDataset
-from .typo_dataset import TypoDataset
-from .typo_inference_dataset import TypoInferenceDataset
-from .word_dataset import WordDataset
-from .word_inference_dataset import WordInferenceDataset
+from .char import CharDataset
+from .char_inference import CharInferenceDataset
+from .typo import TypoDataset
+from .typo_inference import TypoInferenceDataset
+from .word import WordDataset
+from .word_inference import WordInferenceDataset
 
 __all__ = [
     "TypoDataset",
