+++ tests/datamodule/datasets/test_char_dataset.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 from transformers import AutoTokenizer, PreTrainedTokenizerBase
 
-from kwja.datamodule.datasets.char_dataset import CharDataset
+from kwja.datamodule.datasets import CharDataset
 from kwja.utils.constants import IGNORE_INDEX, WORD_NORM_OP_TAGS, WORD_SEGMENTATION_TAGS
